requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-core==0.3.51
langchain-mcp-adapters==0.0.8
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
lxml==5.3.2
markdownify==1.1.0
mcp==1.6.0
mcp-server-fetch==2025.4.7
openai==1.74.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
Protego==0.4.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
readabilipy==0.3.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
webencodings==0.5.1
xxhash==3.5.0
yt-dlp==2025.3.31
zstandard==0.23.0
